move test util file to tests folder
